==> code/01_QC.ipynb <==
# Step 1: Quality Control and pilot selection

import os
import glob
from pathlib import Path
#from dotenv import load_dotenv
import pandas as pd
import numpy as np
# import dask.dataframe as dd
# import dask.bag as db

import matplotlib.pyplot as plt
import seaborn as sns


nat_img_dir='/nese/mit/group/sig/projects/naturalistic/nat_img/'


## load saved qc csvs from MRI-QC

can skip this if re-running and just load the new saved csv

movieDM_qc = pd.read_csv(f"{nat_img_dir}sourcedata/data/HBN/phenotype/movieDM_qc.csv")
movieTP_qc = pd.read_csv(f"{nat_img_dir}sourcedata/data/HBN/phenotype/movieTP_qc.csv")
rest1_qc = pd.read_csv(  f"{nat_img_dir}sourcedata/data/HBN/phenotype/rest1_qc.csv"  )
rest2_qc = pd.read_csv(  f"{nat_img_dir}sourcedata/data/HBN/phenotype/rest2_qc.csv"  )

#combine all of them into one and format
import ast
all_qc = pd.concat([movieDM_qc, movieTP_qc, rest1_qc, rest2_qc])

all_qc['task'] = all_qc['task'].str.replace('run-1', '1')
all_qc['task'] = all_qc['task'].str.replace('run-2', '2')
all_qc['task'] = all_qc['task'].str.replace('task-movieTP', 'movieTP')
all_qc['task'] = all_qc['task'].str.replace('task-movieDM', 'movieDM')
all_qc.rename(columns={'task': 'run_bad'}, inplace=True)

# all_qc['subj_id'] = all_qc['subj_id'].str.replace('sub-', '')
# all_qc.rename(columns={'subj_id': 'sub'}, inplace=True)

# Assuming 'all_qc' is your DataFrame and 'bids_meta' is your column
all_qc['bids_meta'] = all_qc['bids_meta'].apply(ast.literal_eval)
new_columns = all_qc['bids_meta'].apply(pd.Series)
# Now, 'new_columns' is a DataFrame where each key-value pair in the dictionaries
# are separated into columns and values. You can add these new columns to your original DataFrame.
all_qc = pd.concat([all_qc, new_columns[['session_id', 'subject_id', 'task_id']]], axis=1)

all_qc.rename(columns={'task_id': 'task'}, inplace=True)
all_qc.rename(columns={'session_id': 'ses'}, inplace=True)
all_qc.rename(columns={'subject_id': 'sub'}, inplace=True)



#filter them 1) leniently and 2) more stringently and save these 2 filters as binary columns
all_qc['relaxed_fd50'] = np.where(all_qc['fd_perc'] < 50, 1, 0)
all_qc['dorit_fd40'] = np.where(all_qc['fd_perc'] < 40, 1, 0)
#save this file
all_qc.to_csv('../data/all_qc.csv', index=False)

#these subjects were ppotentially erroneously exlcuded before due to incorrect task name

acq=all_qc[all_qc['run_bad'].str.contains('acq')]

# unique_values = acq['run_bad'].unique()
# print(unique_values)
acq

## load fixed qc dataframe

all_qc = pd.read_csv(  f"../data/all_qc.csv"  )

## Load demographic (dx) data

def get_imaging_site(identifier):
    path = f'/nese/mit/group/sig/projects/hbn/hbn_bids/sub-{identifier}/'
    if os.path.isdir(path + 'ses-HBNsiteRU'):
        return 'HBNsiteRU'
    elif os.path.isdir(path + 'ses-HBNsiteCBIC'):
        return 'HBNsiteCBIC'
    elif os.path.isdir(path + 'ses-HBNsiteSI'):
        return 'HBNsiteSI'
    elif os.path.isdir(path + 'ses-HBNsiteCUNY'):
        return 'HBNsiteCUNY'
    else:
        return None
        
dx_final=pd.read_csv(f'{nat_img_dir}sourcedata/data/HBN/phenotype/parsed/dx_onehot.csv')
dx_final['imaging_site'] = dx_final['Identifiers'].apply(get_imaging_site)

filtered_df = dx_final[dx_final['Autism Spectrum Disorder'] == 1]
asd_df = filtered_df[['Identifiers','imaging_site']].rename(columns={'Identifiers': 'subj_id'})
asd_df['subj_id'] = 'sub-' + asd_df['subj_id'].astype(str)

filtered_df = dx_final[dx_final['No Diagnosis Given'] == 1]
nt_df = filtered_df[['Identifiers','imaging_site']].rename(columns={'Identifiers': 'subj_id'})
nt_df['subj_id'] = 'sub-' + nt_df['subj_id'].astype(str)

filtered_df = dx_final[(dx_final['No Diagnosis Given'] == 0) & (dx_final['Autism Spectrum Disorder'] == 0)]
pilot_df = filtered_df[['Identifiers','imaging_site']].rename(columns={'Identifiers': 'subj_id'})
pilot_df['subj_id'] = 'sub-' + pilot_df['subj_id'].astype(str)

dx_final

### filter the subjects by task (dm), site (ru), quality (relaxed), dx for matching

'''
I want a dataframe which has:

from all_qc
- only the relaxed filtered subjects
- only task==DM

from dx_final:
- a 'case' column where ASD (1) and controls (0), others (-1)
- only ses==RU
'''

dx_final_ru = dx_final[dx_final['imaging_site'] == 'HBNsiteRU']

def assign_class(row):
    if row['Autism Spectrum Disorder'] == 1:
        return 1
    elif row['No Diagnosis Given'] == 1:
        return 0
    else:
        return -1

dx_final_ru['class'] = dx_final_ru.apply(assign_class, axis=1)

all_qc_ru_dm = all_qc[all_qc['ses'] == 'HBNsiteRU']
all_qc_ru_dm = all_qc_ru_dm[all_qc_ru_dm['task'] == 'movieDM']
all_qc_ru_dm = all_qc_ru_dm[all_qc_ru_dm['relaxed_fd50'] == 1]

#get the common subjects between these dataframes
dx_final_ru = dx_final_ru[dx_final_ru['Identifiers'].isin(all_qc_ru_dm['sub'])]


### match the ASD to NT by age and sex and then match the matches to non-asd, non-NT for RU pilot subset

# Load the data into a pandas dataframe
df = dx_final_ru[['Identifiers', 'Age', 'Sex', 'class','imaging_site']]

#df = df.loc[(df['imaging_site'] == 'RU') | (df['imaging_site'] == 'CBIC')]


# Separate cases and controls based on the `Group` variable
cases = df[df['class'] == 0]
controls = df[df['class'] == 1]

# Rename the variables for cases and controls to distinguish them
cases = cases.rename(columns={'Age': 'Age_case'})
controls = controls.rename(columns={'Age': 'Age_control'})

# Join the cases and controls based on age and sex
matched = pd.merge(cases, controls, on=['Sex'])
matched = matched[(matched['Age_control'] - 1 <= matched['Age_case']) & (matched['Age_case'] <= matched['Age_control'] + 1)]

# Randomly select one match if there are more
matched = matched.sample(frac=1).groupby('Identifiers_x').head(1)
matched = matched.sample(frac=1).groupby('Identifiers_y').head(1)

###
#Now using these matched asd and controls, make a matched pilot subset from the non-control non-asd population
###


# Separate cases and controls based on the `Group` variable
controls = df[df['class'] == -1]

# Rename the variables for cases and controls to distinguish them
#cases = cases.rename(columns={'Age': 'Age_case'})
cases= matched[['Identifiers_x', 'Age_case', 'Sex', 'class_x','imaging_site_x']]
cases.rename(columns={'Identifiers_x': 'Identifiers'}, inplace=True)
cases.rename(columns={'class_x': 'class'}, inplace=True)
cases.rename(columns={'imaging_site_x': 'imaging_site'}, inplace=True)


controls = controls.rename(columns={'Age': 'Age_control'})

# Join the cases and controls based on age and sex
matched2 = pd.merge(cases, controls, on=['Sex'])
matched2 = matched2[(matched2['Age_control'] - 1 <= matched2['Age_case']) & (matched2['Age_case'] <= matched2['Age_control'] + 1)]

# Randomly select one match if there are more
matched2 = matched2.sample(frac=1).groupby('Identifiers_x').head(1)
matched2 = matched2.sample(frac=1).groupby('Identifiers_y').head(1)

pilots = matched2[['Identifiers_y', 'Age_control', 'Sex', 'class_y','imaging_site_y']]

pilots

# pilot_subjects=pd.read_csv('../data/pilots_ru_dm.csv') # load pilot subjects
import nibabel as nb
subjects=[]
for sub in list(pilots['Identifiers_y']):
    
    im_file = f'/nese/mit/group/sig/projects/hbn/hbn_bids/derivatives/xcp_d_0.7.1/sub-{sub}/ses-HBNsiteRU/func/sub-{sub}_ses-HBNsiteRU_task-movieDM_space-fsLR_seg-Glasser_den-91k_stat-mean_timeseries.ptseries.nii'
    try:
        img = nb.load(im_file)
        img_y = img.get_fdata()
        print(img_y.shape)
        subjects.append(sub)
    except:
        print(f'missing {sub}')
print(f'loaded {len(subjects)} subjects')
#return(subjects)

# pilots.to_csv('../data/pilots_ru_dm.csv', index=False)

# pilots['Identifiers_y'].to_csv('pilots_ru_dm_list.txt', header=False, index=False)

### make a new pilot selection; remove the current pilots from df

# Load the data into a pandas dataframe
df = dx_final_ru[['Identifiers', 'Age', 'Sex', 'class','imaging_site']]



with open('pilots_ru_dm_list.txt', 'r') as file:
    file_list = [line.strip() for line in file]


df = dx_final_ru[~dx_final_ru['Identifiers'].isin(file_list)]




#df = df.loc[(df['imaging_site'] == 'RU') | (df['imaging_site'] == 'CBIC')]


# Separate cases and controls based on the `Group` variable
cases = df[df['class'] == 0]
controls = df[df['class'] == 1]

# Rename the variables for cases and controls to distinguish them
cases = cases.rename(columns={'Age': 'Age_case'})
controls = controls.rename(columns={'Age': 'Age_control'})

# Join the cases and controls based on age and sex
matched = pd.merge(cases, controls, on=['Sex'])
matched = matched[(matched['Age_control'] - 1 <= matched['Age_case']) & (matched['Age_case'] <= matched['Age_control'] + 1)]

# Randomly select one match if there are more
matched = matched.sample(frac=1).groupby('Identifiers_x').head(1)
matched = matched.sample(frac=1).groupby('Identifiers_y').head(1)

###
#Now using these matched asd and controls, make a matched pilot subset from the non-control non-asd population
###


# Separate cases and controls based on the `Group` variable
controls = df[df['class'] == -1]

# Rename the variables for cases and controls to distinguish them
#cases = cases.rename(columns={'Age': 'Age_case'})
cases= matched[['Identifiers_x', 'Age_case', 'Sex', 'class_x','imaging_site_x']]
cases.rename(columns={'Identifiers_x': 'Identifiers'}, inplace=True)
cases.rename(columns={'class_x': 'class'}, inplace=True)
cases.rename(columns={'imaging_site_x': 'imaging_site'}, inplace=True)


controls = controls.rename(columns={'Age': 'Age_control'})

# Join the cases and controls based on age and sex
matched2 = pd.merge(cases, controls, on=['Sex'])
matched2 = matched2[(matched2['Age_control'] - 1 <= matched2['Age_case']) & (matched2['Age_case'] <= matched2['Age_control'] + 1)]

# Randomly select one match if there are more
matched2 = matched2.sample(frac=1).groupby('Identifiers_x').head(1)
matched2 = matched2.sample(frac=1).groupby('Identifiers_y').head(1)

pilots = matched2[['Identifiers_y', 'Age_control', 'Sex', 'class_y','imaging_site_y']]

pilots.to_csv('../data/pilots2_ru_dm.csv', index=False)
pilots['Identifiers_y'].to_csv('pilots2_ru_dm_list.txt', header=False, index=False)
