==> code/03-video_feature_extraction.ipynb <==
# Installing Net2Brain

#use conda env net2brain

#!pip install -U git+https://github.com/cvai-roig-lab/Net2Brain

# Step 1: Feature Extraction

## Using `FeatureExtractor` with a model from Net2Brain

The FeatureExtractor class provides an interface for extracting features from a given model. When initializing this class, you can customize its behavior by setting various parameters:

- `model` (required): The model from which you want to extract features. Either string in combination with a netset (next parameter), or a variable with a model-type.
- `netset` (optional): The netset (collection of networks) that the model belongs to.
- `layers_to_extract` (optional): A list of layer names or indices from which you want to extract features. Default is None, indicating that all layers preset in the toolbox will be used.
- `device` (optional): The device on which to perform the computations, e.g., 'cuda' for GPU or 'cpu' for CPU. Default is None, which will use the device specified in the global PyTorch settings.
- `transforms` (optional): A list of data preprocessing transforms to apply to the input data before passing it through the model. Default is None, which uses the preset transformations
- `pretrained` (optional): A boolean flag indicating whether to use a pretrained model (if available) or to initialize the model with random weights. Default is True, which means that a pretrained model will be used if possible.

- - -


First we need to a dataset to play around with. For that we will use the dataset by [Micheal F. Bonner (2017)](https://www.pnas.org/doi/full/10.1073/pnas.1618228114), which we can download using the `load_dataset` function

# from net2brain.utils.download_datasets import load_dataset
# stimuli_path, roi_path = load_dataset("bonner_pnas2017")

### Initating FeatureExtractor


To extract the activations of a pretrained model from a netset, you can use the FeatureExtractor class. First, you need to initialize the class by providing the name of the model and the name of the netset. You can find a suitable model and netset by exploring the taxonomy options available in the Net2Brain toolbox, as shown in the previous notebook "0_Exploring_Net2Brain". For instance, in the following example, we will use AlexNet from the standard netset.

from net2brain.feature_extraction import FeatureExtractor
fx = FeatureExtractor(model='ResNet50', netset='Standard', device='cpu') #cuda not working now, try dif gpu?

The `extract` method computes feature extraction from an image dataset. It takes in the following parameters:

- `data_path` (required): The path to the images from which to extract the features. The images must be in JPEG or PNG format.
- `save_path` (optional): The path to save the extracted features to. If None, the folder where the features are saved is named after the current date in the format "{year}{month}{day}{hour}{minute}".
- `layers_to_extract` (optional): A list of layer names or indices from which to extract features. If None, the specified layers will be used.
- `consolidate_per_layer` (optional): The features are extracted image-wise. This is defaulted to true and will consolidate them per layer if not set to False. Defautls to True.
- `dim_reduction` (optional): Type of dimensionality reduction (For now: SRP) for extracted features. Defaults to None.
- `n_components` (optonal): Number of components for dimensionality reduction. Defaults to 50.

from net2brain.feature_extraction import FeatureExtractor
import os

stim='DM'
stimuli_path = f'../data/{stim}_frames/'
save_path = f'../data/{stim}_frames_resnet50/'

if not os.path.isdir(save_path):
    os.mkdir(save_path) 
    
#relu, maxpool, block1, block2, block3, block4, avgpool
layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']

#fx = FeatureExtractor(model='AlexNet', netset='Standard', device='cpu')
fx.extract(data_path=stimuli_path, save_path=save_path, layers_to_extract=layers_to_extract)  #, consolidate_per_layer=False)

from net2brain.feature_extraction import FeatureExtractor
import os

for stim in ['friends_s01e01a','friends_s01e01b','friends_s01e02a','friends_s01e02b']:
    stimuli_path = f'../data/{stim}_frames/'
    save_path = f'../data/{stim}_frames_resnet50/'
    
    if not os.path.isdir(save_path):
        os.mkdir(save_path) 
        
    #relu, maxpool, block1, block2, block3, block4, avgpool
    layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']
    
    #fx = FeatureExtractor(model='AlexNet', netset='Standard', device='cpu')
    fx.extract(data_path=stimuli_path, save_path=save_path, layers_to_extract=layers_to_extract)  #, consolidate_per_layer=False)

__Net2Brain__ chooses by default from which layers of the model to extract the features from. You can inspect which layers are selected by default by calling the `layers_to_extract` attribute:

fx.layers_to_extract

These are not all the layers that **can** be extracted. If you want to see all the layers that can possibly be extracted you you call `get_all_layers()`.

fx.get_all_layers()

If you wish to change the layers to be extracted you can add it to the `extract` function like with the parameter 
```
fx.extract(..., layers_to_extract=[your_layers])
```

## explore features

import os
import numpy as np
import glob

stim='DM'
save_path = f'../data/{stim}_frames_resnet50/'

#print the
emb_list = glob.glob(f'{save_path}*.npz')

layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']

print('ResNet50 frame embeddings')
# for emb_f in emb_list:
#     emb = np.load(f'{emb_f}')
#     print(emb['frame_0000.jpg'].shape, os.path.splitext(os.path.basename(emb_f))[0])
for layer in layers_to_extract:
    emb = np.load(f'{save_path}{layer}.npz')
    print(emb['frame_0000.jpg'].shape, layer)


emb['frame_0000.jpg'].shape

## try dimensionality reduction features, what shape are they?

from net2brain.feature_extraction import FeatureExtractor
fx = FeatureExtractor(model='ResNet50', netset='Standard', device='cpu') #cuda not working now, try dif gpu?
from net2brain.feature_extraction import FeatureExtractor
import os

stim='DM'
stimuli_path = f'../data/{stim}_frames/'
save_path = f'../data/{stim}_frames_resnet50_srp_50/'

if not os.path.isdir(save_path):
    os.mkdir(save_path) 
    
#relu, maxpool, block1, block2, block3, block4, avgpool
layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']

#fx = FeatureExtractor(model='AlexNet', netset='Standard', device='cpu')
fx.extract(data_path=stimuli_path, save_path=save_path, layers_to_extract=layers_to_extract, dim_reduction='srp', n_components=50)  #, consolidate_per_layer=False)

import os
import numpy as np
import glob

stim='DM'
save_path = f'../data/{stim}_frames_resnet50_srp_50/'

#print the
emb_list = glob.glob(f'{save_path}*.npz')

layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']

print('ResNet50 frame embeddings')
# for emb_f in emb_list:
#     emb = np.load(f'{emb_f}')
#     print(emb['frame_0000.jpg'].shape, os.path.splitext(os.path.basename(emb_f))[0])
for layer in layers_to_extract:
    emb = np.load(f'{save_path}{layer}.npz')
    print(emb['frame_0000.jpg'].shape, layer)


### hOW DO THEY DO THIS???
it seems like they apply dimensionality reduction to each image (features) individually... how do they end up with exactly 50

emb['frame_0000.jpg']

import numpy as np
from sklearn.decomposition import PCA

# Generate a random array of size (1, 2000)
data = np.random.rand(2, 2000)

# Initialize PCA
pca = PCA(n_components=2)

# Fit and transform the data
pca_result = pca.fit_transform(data)

print("PCA Result:\n", pca_result.shape)


## now try to get some additional features! like actual video features
#use conda env net2brain2

from net2brain.feature_extraction import FeatureExtractor
fx = FeatureExtractor(model='slowfast_r50', netset='Pyvideo', device='cuda') #cuda worked with 1080ti

#fx.get_all_layers()
fx.layers_to_extract

from net2brain.feature_extraction import FeatureExtractor
import os

stim='DM'
stimuli_path = f'../data/{stim}_videos/'
save_path = f'../data/{stim}_videos_slowfast_r50/'

if not os.path.isdir(save_path):
    os.mkdir(save_path) 
    
#relu, maxpool, block1, block2, block3, block4, avgpool
#layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']

#fx = FeatureExtractor(model='AlexNet', netset='Standard', device='cpu')
fx.extract(data_path=stimuli_path, save_path=save_path)  #, consolidate_per_layer=False)

## try resnet50 with video inputs so they average - maybe better?

from net2brain.feature_extraction import FeatureExtractor
fx = FeatureExtractor(model='ResNet50', netset='Standard', device='cuda') #cuda not working now, try dif gpu?

import os

stim='DM'
stimuli_path = f'../data/{stim}_videos/'
save_path = f'../data/{stim}_videos_resnet50/'

if not os.path.isdir(save_path):
    os.mkdir(save_path) 
    
#relu, maxpool, block1, block2, block3, block4, avgpool
layers_to_extract=['relu','maxpool', 'layer1', 'layer2', 'layer3', 'layer4', 'avgpool']

#fx = FeatureExtractor(model='AlexNet', netset='Standard', device='cpu')
fx.extract(data_path=stimuli_path, save_path=save_path, layers_to_extract=layers_to_extract)  #, consolidate_per_layer=False)

## extract motion features using pymoten

for pymoten, use hbn_asd conda env

#### Despicable Me

import moten

import cv2

# Path to the video file
video_path = '../data/DM.mp4'

# Open the video file
video = cv2.VideoCapture(video_path)

# Get the total number of frames
total_frames = int(video.get(cv2.CAP_PROP_FRAME_COUNT))

print(f'Total number of frames: {total_frames}')

width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))

print(f'Width: {width}, Height: {height}')

fps = video.get(cv2.CAP_PROP_FPS)

print(f'FPS: {fps}')

# Release the video capture object
video.release()

video_file = '../data/DM.mp4'
small_vhsize = (72, 108)        # height x width
#luminance_images = moten.io.video2luminance(video_file, size=small_vhsize, nimages=100)
luminance_images = moten.io.video2luminance(video_file, size=small_vhsize)
# luminance_images.shape
# plt.imshow(luminance_images[0,:,:])
# plt.imshow(luminance_images[99,:,:])

# eta 4 min
# Create a pyramid of spatio-temporal gabor filters
nimages, vdim, hdim = luminance_images.shape
pyramid = moten.get_default_pyramid(vhsize=(vdim, hdim), fps=30)

# Compute motion energy features
moten_features = pyramid.project_stimulus(luminance_images)

import h5py
# Path to the HDF5 file
hdf5_path = '../data/features/DM_pymoten.h5'

# Create an HDF5 file
with h5py.File(hdf5_path, 'w') as hdf5_file:
    # Create a dataset in the file
    hdf5_file.create_dataset('pymoten', data=moten_features)

print(f'Motion features saved to {hdf5_path}')


# # Path to the HDF5 file
# hdf5_path = '../data/features/DM_pymoten.h5'

# # Open the HDF5 file
# with h5py.File(hdf5_path, 'r') as hdf5_file:
#     # Access the dataset
#     motion_features = hdf5_file['pymoten'][:]

# # Print the shape of the loaded data
# print(f'Motion features shape: {motion_features.shape}')


#### The Present

import cv2
# Path to the video file
video_path = '../data/TP.mp4'
# Open the video file
video = cv2.VideoCapture(video_path)
# Get the total number of frames
total_frames = int(video.get(cv2.CAP_PROP_FRAME_COUNT))
print(f'Total number of frames: {total_frames}')
width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
print(f'Width: {width}, Height: {height}')
fps = video.get(cv2.CAP_PROP_FPS)
print(f'FPS: {fps}')
# Release the video capture object
video.release()

stim='TP'
video_file = f'../data/{stim}.mp4'
small_vhsize = (72, 128)        # height x width
#luminance_images = moten.io.video2luminance(video_file, size=small_vhsize, nimages=100)
luminance_images = moten.io.video2luminance(video_file, size=small_vhsize)
# luminance_images.shape
# plt.imshow(luminance_images[0,:,:])
# plt.imshow(luminance_images[99,:,:])
# eta 4 min
# Create a pyramid of spatio-temporal gabor filters
nimages, vdim, hdim = luminance_images.shape
pyramid = moten.get_default_pyramid(vhsize=(vdim, hdim), fps=24)

# Compute motion energy features
moten_features = pyramid.project_stimulus(luminance_images)

import h5py
# Path to the HDF5 file
hdf5_path = f'../data/features/{stim}_pymoten.h5'

# Create an HDF5 file
with h5py.File(hdf5_path, 'w') as hdf5_file:
    # Create a dataset in the file
    hdf5_file.create_dataset('pymoten', data=moten_features)

print(f'Motion features saved to {hdf5_path}')


#### Friends

import cv2
# Path to the video file
stim='friends_s01e02a'
video_path = f'/om2/scratch/tmp/jsmentch/nat_img/sourcedata/data/cneuromod/friends.stimuli/s1/{stim}.mkv'
# Open the video file
video = cv2.VideoCapture(video_path)
# Get the total number of frames
total_frames = int(video.get(cv2.CAP_PROP_FRAME_COUNT))
print(f'Total number of frames: {total_frames}')
width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
print(f'Width: {width}, Height: {height}')
fps = video.get(cv2.CAP_PROP_FPS)
print(f'FPS: {fps}')
# Release the video capture object
video.release()

video_file = video_path
small_vhsize = (72, 108)        # height x width
#luminance_images = moten.io.video2luminance(video_file, size=small_vhsize, nimages=100)
luminance_images = moten.io.video2luminance(video_file, size=small_vhsize)
# luminance_images.shape
# plt.imshow(luminance_images[0,:,:])
# plt.imshow(luminance_images[99,:,:])
# eta 4 min
# Create a pyramid of spatio-temporal gabor filters
nimages, vdim, hdim = luminance_images.shape
pyramid = moten.get_default_pyramid(vhsize=(vdim, hdim), fps=30)

# Compute motion energy features
moten_features = pyramid.project_stimulus(luminance_images)

import h5py
# Path to the HDF5 file
hdf5_path = f'../data/features/{stim}_pymoten.h5'

# Create an HDF5 file
with h5py.File(hdf5_path, 'w') as hdf5_file:
    # Create a dataset in the file
    hdf5_file.create_dataset('pymoten', data=moten_features)

print(f'Motion features saved to {hdf5_path}')

import cv2
stim='friends_s01e02b'

# Path to the video file
video_path = f'/om2/scratch/tmp/jsmentch/nat_img/sourcedata/data/cneuromod/friends.stimuli/s1/{stim}.mkv'
# Open the video file
video = cv2.VideoCapture(video_path)
# Get the total number of frames
total_frames = int(video.get(cv2.CAP_PROP_FRAME_COUNT))
print(f'Total number of frames: {total_frames}')
width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
print(f'Width: {width}, Height: {height}')
fps = video.get(cv2.CAP_PROP_FPS)
print(f'FPS: {fps}')
# Release the video capture object
video.release()

video_file = video_path
small_vhsize = (72, 108)        # height x width
#luminance_images = moten.io.video2luminance(video_file, size=small_vhsize, nimages=100)
luminance_images = moten.io.video2luminance(video_file, size=small_vhsize)
# luminance_images.shape
# plt.imshow(luminance_images[0,:,:])
# plt.imshow(luminance_images[99,:,:])
# eta 4 min
# Create a pyramid of spatio-temporal gabor filters
nimages, vdim, hdim = luminance_images.shape
pyramid = moten.get_default_pyramid(vhsize=(vdim, hdim), fps=30)

# Compute motion energy features
moten_features = pyramid.project_stimulus(luminance_images)

import h5py
# Path to the HDF5 file
hdf5_path = f'../data/features/{stim}_pymoten.h5'

# Create an HDF5 file
with h5py.File(hdf5_path, 'w') as hdf5_file:
    # Create a dataset in the file
    hdf5_file.create_dataset('pymoten', data=moten_features)

print(f'Motion features saved to {hdf5_path}')
