==> code/19_mean_predictions.ipynb <==
# notebook to test the mean_prediction methods

for the R2 evaluation can you compare to predicting the mean signal? so replace the predicted time course by its mean during training and see how the R2's compare. and you can also run this evaluation simply by computing the mean of the test block and asking if you could predict the mean how well would you do? we may consider that as a lower bound on R2. 



I'm trying to work this out but I need a little clarification since there seem to be a few different ways to do this

My interpretation is for each subject, for each grayordinate in the ROI: see how well the mean of the training timecourse of that grayordinate predicts (using R2) the test block of that grayordinate. The predictions for each time point of each fold (the same for each time point but different across folds) are the mean of the corresponding training timepoints. I can compute the R2 and get one value for each grayordinate in the ROI and then I can take the mean of those. Is that right? Or since you are saying:

"asking if you could predict the mean how well would you do?"

I can use the mean of the training timepoints to predict the mean of the testing timepoints and calculate R2 over the cv folds or over the ROI. Or simply the mean training ROI timecourse (all grayordinates averaged to one time course) to predict the mean testing ROI timecourse.

#### TODOs:

for 8 good subjects:

1) How well does the mean (of training timecourse, each voxel has a mean value) of the training predict the real testing timecourse

$R^2$ all zeros/negative

2) How well does the mean (of training timecourse, each voxel has a mean value) of the training predict the mean (of testing timecourse, each voxel has a mean value) testing voxels

$R^2$ all zeros/negative

4) How well is the mean (over the timecourse and the roi) predicted

$R^2$ all zeros/negative

5) within each CV fold, calculate $R^2$ (across grayordinates) of the mean of the training to the mean of the testing

$R^2$ all zeros/negative


with open('good_pilots_ru_dm_list.txt', 'r') as file:
    sub_list = file.readlines()
good_subjects_list = [line.strip() for line in sub_list]

#for sub in good_subjects_list[:3]:
sub='NDARHJ830RXD'
%run pilot.py -s $sub -p a4a5 -f concatspeech -d 0 -m

for sub in good_subjects_list:
#sub='NDARHJ830RXD'
    %run pilot.py -s $sub -p a4a5 -f concatspeech -d 0 -m
