==> code/28_NEW_QC.ipynb <==
import pandas as pd
import os
import glob
from pathlib import Path
#from dotenv import load_dotenv
import numpy as np
# import dask.dataframe as dd
# import dask.bag as db

import matplotlib.pyplot as plt
import seaborn as sns

## generate merged csv with all of the data demographics, diagnoses etc

nat_img_dir='/nese/mit/group/sig/projects/naturalistic/nat_img/'

all_qc = pd.read_csv(  f"../data/all_qc.csv"  )
all_qc = all_qc[all_qc['task'].isin(['movieDM', 'movieTP'])]

def get_imaging_site(identifier):
    path = f'/nese/mit/group/sig/projects/hbn/hbn_bids/sub-{identifier}/'
    if os.path.isdir(path + 'ses-HBNsiteRU'):
        return 'HBNsiteRU'
    elif os.path.isdir(path + 'ses-HBNsiteCBIC'):
        return 'HBNsiteCBIC'
    elif os.path.isdir(path + 'ses-HBNsiteSI'):
        return 'HBNsiteSI'
    elif os.path.isdir(path + 'ses-HBNsiteCUNY'):
        return 'HBNsiteCUNY'
    else:
        return None
        
dx_final=pd.read_csv(f'{nat_img_dir}sourcedata/data/HBN/phenotype/parsed/dx_onehot.csv')
dx_final['imaging_site'] = dx_final['Identifiers'].apply(get_imaging_site)

filtered_df = dx_final[dx_final['Autism Spectrum Disorder'] == 1]
asd_df = filtered_df[['Identifiers','imaging_site']].rename(columns={'Identifiers': 'subj_id'})
asd_df['subj_id'] = 'sub-' + asd_df['subj_id'].astype(str)

filtered_df = dx_final[dx_final['No Diagnosis Given'] == 1]
nt_df = filtered_df[['Identifiers','imaging_site']].rename(columns={'Identifiers': 'subj_id'})
nt_df['subj_id'] = 'sub-' + nt_df['subj_id'].astype(str)

filtered_df = dx_final[(dx_final['No Diagnosis Given'] == 0) & (dx_final['Autism Spectrum Disorder'] == 0)]
pilot_df = filtered_df[['Identifiers','imaging_site']].rename(columns={'Identifiers': 'subj_id'})
pilot_df['subj_id'] = 'sub-' + pilot_df['subj_id'].astype(str)

merged_qc = all_qc.merge(dx_final, left_on='sub', right_on='Identifiers', how='inner')

import pandas as pd

# Initialize an empty list to store the max FD values
max_fd_values = []

for index, row in merged_qc.iterrows():
    try:
        # Generate the confounds_file path based on the row's sub, ses, and task columns
        confounds_file = f'/nese/mit/group/sig/projects/hbn/hbn_bids/derivatives/fmriprep_23.2.0/sub-{row["sub"]}/ses-{row["ses"]}/func/sub-{row["sub"]}_ses-{row["ses"]}_task-{row["task"]}_desc-confounds_timeseries.tsv'
        
        # Load the confounds file and extract the max FD value
        confounds_df = pd.read_csv(confounds_file, sep='\t')
        max_fd_value = confounds_df['framewise_displacement'].max()
        #print(max_fd_value)
        # Append the max FD value to the list
        max_fd_values.append(max_fd_value)
    except:
        max_fd_values.append('nan')
# Assign the list as a new column in the DataFrame
merged_qc['max_fd'] = max_fd_values


(merged_qc['max_fd'] == 'nan').sum()

690 have 'nan' so they do not have the confounds file available... UH DOUBLE CHECK THIS WHAT ABT THE ACQ VARIANT PPL

#make a column for if the t1 exists
t1_exists = []

for index, row in merged_qc.iterrows():
    # Generate the confounds_file path based on the row's sub, ses, and task columns
    t1_file=f'/nese/mit/group/sig/projects/hbn/hbn_bids/sub-{row["sub"]}/ses-{row["ses"]}/anat/sub-{row["sub"]}_ses-{row["ses"]}_*T1w.nii.gz'
    matching_files = glob.glob(t1_file)

    if matching_files:
        #print(f"{t1_file} exists")
        t1_exists.append(1)
    else:
        #print(f"{t1_file} does not exist")
        t1_exists.append(0)
# Assign the list as a new column in the DataFrame
merged_qc['t1_exists'] = t1_exists

import json
#make a column for the kwyh uncertainty
uncertainty = []

for index, row in merged_qc.iterrows():
    # Generate the confounds_file path based on the row's sub, ses, and task columns
    #t1_file=f'/nese/mit/group/sig/projects/hbn/hbn_bids/sub-{row["sub"]}/ses-{row["ses"]}/anat/sub-{row["sub"]}_ses-{row["ses"]}_*T1w.nii.gz'
    try:
        uncertainty_file=f'/nese/mit/group/sig/projects/hbn/hbn_bids/derivatives/kwyk/{row["sub"]}_uncertainty.json'
        with open(uncertainty_file, 'r') as file:
            data = json.load(file)
        uncertainty.append(data['uncertainty'])
    except:
        #print(f"{t1_file} does not exist")
        uncertainty.append('nan')
# Assign the list as a new column in the DataFrame
merged_qc['uncertainty'] = uncertainty

(merged_qc['uncertainty'] == 'nan').sum()

uncertainty_masked_list=[]
for index, row in merged_qc.iterrows():
    # Generate the confounds_file path based on the row's sub, ses, and task columns
    #t1_file=f'/nese/mit/group/sig/projects/hbn/hbn_bids/sub-{row["sub"]}/ses-{row["ses"]}/anat/sub-{row["sub"]}_ses-{row["ses"]}_*T1w.nii.gz'
    try:
        uncertainty_file=f'/nese/mit/group/sig/projects/hbn/hbn_bids/derivatives/kwyk/{row["sub"]}_uncertainty_masked.txt'
        with open(uncertainty_file, 'r') as file:
            number = float(file.read().strip())
        uncertainty_masked_list.append(number)
    except:
        #print(f"{t1_file} does not exist")
        uncertainty_masked_list.append('nan')
merged_qc['uncertainty_masked'] = uncertainty_masked_list


merged_qc.to_csv('../data/merged_qc.csv', index=False)

## do stuff with the merged_qc data

merged_qc = pd.read_csv('../data/merged_qc.csv', index_col=False)

pd.set_option('display.max_columns', None)
print(merged_qc)

### how many subjects are lost to no existing t1?

other_qc = merged_qc[(merged_qc['No Diagnosis Given'] == 0) & (merged_qc['Autism Spectrum Disorder'] == 0)]
other_qc_DM=other_qc[(other_qc['task']=='movieDM')]
other_qc_TP=other_qc[(other_qc['task']=='movieTP')]

asd_qc = merged_qc[(merged_qc['Autism Spectrum Disorder'] == 1)]
asd_qc_DM=asd_qc[(asd_qc['task']=='movieDM')]
asd_qc_TP=asd_qc[(asd_qc['task']=='movieTP')]

control_qc = merged_qc[(merged_qc['No Diagnosis Given'] == 1)]
control_qc_DM=control_qc[(control_qc['task']=='movieDM')]
control_qc_TP=control_qc[(control_qc['task']=='movieTP')]


#total subjects

for df in [merged_qc, asd_qc, control_qc, other_qc]:
#    print(f'{df}')
    df_name = [name for name in globals() if globals()[name] is df][0]
    print(f"{df_name}")

    print( 'all listed:        ',len(df))
    print( 'with t1:           ', len(df[(df['t1_exists'] == 1)]) )
    print( 'max FD <2.4:       ', df[df['max_fd'] < 2.4].shape[0])
    print( 'uncertainty <=0.4: ', df[df['uncertainty'] <= 0.4 ].shape[0])

    print( 'filtered by all:   ', len(df[(df['t1_exists'] == 1) & (df['max_fd'] < 2.4) & (df['uncertainty'] <= 0.4)]))

    print( '\n')


unique_sites = merged_qc['ses'].unique()
print(unique_sites)

leest=['asd','asd_DM','asd_TP',
'control','control_DM','control_TP',
'other','other_DM','other_TP',]
for i,df in enumerate([asd_qc, asd_qc_DM, asd_qc_TP, control_qc, control_qc_DM, control_qc_TP, other_qc, other_qc_DM, other_qc_TP]):
#    print(f'{df}')
    
    print(f'{leest[i]}')
    print( 'all listed:  ',len(df))
    print( 'with t1:     ', len(df[(df['t1_exists'] == 1)]) )
    print( 'max FD <2.4: ', df[df['max_fd'] < 2.4].shape[0])
    print( '\n')


### pick another pilot cohort

#pick 30 new random subjects
with open('pilots_ru_dm_list.txt', 'r') as file:
    exclude_list = file.read().splitlines()

# Step 2: Filter the DataFrame to exclude rows where 'sub' is in the exclude_list
filtered_df = other_qc[~other_qc['sub'].isin(exclude_list)]

# Step 3: Select 26 random rows from the filtered DataFrame
random_rows = filtered_df.sample(n=26)

# Display the selected random rows
sub_column = random_rows['sub']

# Step 5: Save the 'sub' column to a new text file
sub_column.to_csv('pilots_ru_dm_list2.txt', index=False, header=False)

# Verify the output by printing the saved list (optional)
print(sub_column)
