==> code/31_kwyk_uncertainty.ipynb <==
# Notebook to look into the kwyk uncertainty of T1 images across the HBN dataset

import nilearn as nl
from nilearn import plotting, image, masking
import json
import numpy as np
import nibabel as nb
import pandas as pd
import glob
import os
import json

## COUNT how many a) overall b) in each class have uncertainty > 0.4

# list of all subjects 

merged_qc = pd.read_csv('../data/merged_qc.csv', index_col=False)

### get the uncertainty value for each subject

#for df in [merged_qc, asd_qc, control_qc, other_qc]:
df=merged_qc
#    print(f'{df}')
#df_name = [name for name in globals() if globals()[name] is df][0]
#print(f"{df_name}")

print( 'all listed:  ',len(df))
print( 'with t1:     ', len(df[(df['t1_exists'] == 1)]) )
print( 'max FD <2.4: ', df[df['max_fd'] < 2.4].shape[0])
print( '\n')


# first make a function to plot

## first look at pilots

## next look at ASD, Controls
