==> code/32_motion phenotypes.ipynb <==
# motion phenotype investigation-- is it linked to SRS score in pilots?

### first run on pilot1
with open('pilots_ru_dm_list.txt', 'r') as file:
    sub_list = file.readlines()
good_subjects_list = [line.strip() for line in sub_list]

good_subjects_list.remove('NDARZK732FZ0')

#load fd timecourses

#load crown regressor timecourses

#load SRS

#apply umap to them

#color the plot by SRS

### next run on pilot 2
